v17 Phase 4: re-target Bind/Rotate CPI from tag 32 to tag 65 (collision row 43)
- src/cpi.rs: replace TAG_UPDATE_AUTHORITY(32)/AUTHORITY_INSURANCE(2)/34-byte
  wire with TAG_UPDATE_ASSET_AUTHORITY(65)/ASSET_AUTH_INSURANCE(1)/ASSET_INDEX_ZERO
  +2-byte prefix / 36-byte wire per V17 auth overhaul; three footguns pinned by
  byte-level canary tests (tag 32→65, kind 2→1, +asset_index prefix).
  Both cpi_bind_insurance_authority and cpi_rotate_insurance_authority updated.
  cpi_top_up_insurance unchanged (tag 9 / u128 wire is correct for v17).
- src/processor.rs: fix pre-existing clippy unnecessary_cast (amount as u64).
- tests/cpi_tags.rs: rewrite with v17 wire canaries; 5 tests all GREEN.
- tests/v16_stake_insurance_e2e.rs: fix pre-existing clippy warnings (doc list
  indentation, dead_code mint, needless borrows); no logic changes.

Wire: [65u8][0x00 0x00][0x01][pubkey:32] = 36 bytes.
Verified at byte level against v16_program.rs:9407-9464 (handle_update_asset_authority).
Clippy -D warnings: CLEAN. cargo test --lib: 111/111 pass.
Kani: deferred to Phase 5 (wire-correctness is structural, not arithmetic).

+++ src/processor.rs
@@ -1991,7 +1991,7 @@ fn process_return_insurance(
 
     // Validate amount doesn't exceed outstanding insurance
     let outstanding = pool.total_flushed.saturating_sub(pool.total_returned);
-    if (amount as u64) > outstanding {
+    if amount > outstanding {
         msg!(
             "ReturnInsurance: amount {} exceeds outstanding insurance {}",
             amount,
